Fixed issue where compiling quickfix with python and ruby support failed.
 ./configure -v --prefix=~/exchange/exchrel --with-python=~/exchange/exchrel/include --with-ruby

Using the gcc 4.9 compiler everything builds and installs.  It seems that gcc 4.9 is picking up that std::tr1::shared_ptr is available
but that std::shared_ptr is not available.  If I force environment to use -std=c++11 right before the config then the environment for
using std::shared_ptr will get setup and used.  Note that compiling didn't work for me after forcing the ./configure to set the define so it uses std::shared_ptr.