`removeField` generates invalid message in some cases
When removing a field that has the count of repeated group members, the
message will have wrong values for bodylength and checksum.

This happens because there is an inconsistency between calculateString
and calculateLength/calculateChecksum API. calculateString ignores the
repeating group which does not have the counter field while the latter
includes them.

Sample code to demonstrate the problem

https://gist.github.com/balkierode/43759bee0fbbbe5af9f1a964dde93075