NoSides values in FixValues.h are int not char
NoSides is the only NUMINGROUP field that has enum values. The
code generator was generating chars for those values instead of ints.
This commit causes the generator to use int instead of char if the type
is NUMINGROUP (again, only affects NoSides currently).

Also, the sh version of the c++ generate script was not actually generating
FixValues.h and FixTValues.h (but the bat version was).