Refactor and modernize C++ log classes (#568)
* Refactor and modernize C++ log classes

This commit modernizes the C++ logging classes by making use of the 'default' keyword for destructors, adopting the explicit keyword, improving the precision of the 'override' keyword, and optimizing memory usage through the use of 'std::move()'. The addition of preprocessor definitions in 'config.h' further improves the code's platform compatibility.

* Add detailed docstrings to Log classes methods

Introduced meticulous docstrings to the methods in the Log classes that precisely describe their functionalities. Improved the clarity of the code, benefitting developers and contributing to better code maintenance and understanding. The docstrings cover the method's purpose, the explanation for parameters, return types, and the conditions under which the method operates.

* Add detailed docstrings to Log classes methods

Introduced meticulous docstrings to the methods in the Log classes that precisely describe their functionalities. Improved the clarity of the code, benefitting developers and contributing to better code maintenance and understanding. The docstrings cover the method's purpose, the explanation for parameters, return types, and the conditions under which the method operates.

* Add detailed docstrings to Log classes methods

Introduced meticulous docstrings to the methods in the Log classes that precisely describe their functionalities. Improved the clarity of the code, benefitting developers and contributing to better code maintenance and understanding. The docstrings cover the method's purpose, the explanation for parameters, return types, and the conditions under which the method operates.

* Add detailed docstrings to Log classes methods

Introduced meticulous docstrings to the methods in the Log classes that precisely describe their functionalities. Improved the clarity of the code, benefitting developers and contributing to better code maintenance and understanding. The docstrings cover the method's purpose, the explanation for parameters, return types, and the conditions under which the method operates.